requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.2.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
distlib==0.3.8
Django==5.2.7
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.14.0
frozenlist==1.5.0
gevent==25.8.2
greenlet==3.2.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.29.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
multidict==6.1.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pipenv==2023.12.1
platformdirs==4.2.2
prompt-toolkit==3.0.43
propcache==0.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.5
python-telegram-bot-django-persistence==0.2.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.2
traitlets==5.14.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.26.2
watchdog==5.0.0
wcwidth==0.2.13
yarl==1.18.3
zope.event==5.0
zope.interface==6.3
